refactor: 코드 리뷰 피드백 반영 - 재시도 로직, 컨트롤러-서비스 연결, 보고서 개선
## 주요 변경사항

### 1. 재시도 로직 개선
- OrderService: 수동 while문 재시도 → @Retryable로 변경
- ProductService: decreaseStock, increaseStock 메서드에 @Retryable 적용
- PointService: maxAttempts 50 → 10으로 조정 (응답시간 고려)
- 재시도 전략: backoff delay 50-200ms로 통일

### 2. Controller-Service 연결
- 모든 Controller에서 mock response 제거
- Service 계층에서 Response DTO 직접 반환하는 구조로 통일
- CartController, ProductController, CouponController, PaymentController, OrderController 모두 실제 서비스 연결

### 3. Service 메서드 추가
- CartService: getCartItemsWithDetails, addCartItemWithResponse
- ProductService: getProductsWithPaging, getProductDetail, getPopularProductsResponse
- CouponService: issueCouponWithResponse, getUserCouponsWithDetails
- PaymentService: executePaymentWithResponse (@Transactional 추가)
- OrderService: createOrderWithResponse, createOrderFromEntireCart, convertToOrderItemDtos (중복 제거)

### 4. LockManager 제거
- 사용하지 않는 LockManager 클래스 삭제
- OrderService, CouponService에서 의존성 제거

### 5. 보고서 개선 (STEP09)
- MySQL 격리 수준 정정: REPEATABLE_READ가 기본값
- 낙관적 락 데드락 설명 보완: UPDATE 시 배타락 사용, 여러 엔티티 업데이트 시 발생 가능
- 쿠폰 재시도 전략 설명 개선: 재시도와 소진 여부 무관함 명시
- UserCoupon 중복 방지: unique index로 save 시점에도 방지됨 명시
- 포인트 충전 재시도 횟수 조정: 50회 → 10회, 과도한 재시도의 문제점 설명
- 표현 개선: "100% 보장" → "낙관적 락으로 개선", 개선율 단위 명확화

### 6. 주문 기능 확장
- CreateOrderFromCartRequest 추가: 장바구니 전체 주문 API
- CreateOrderRequest: cartItemIds 필드 추가 (일부 선택 주문)
- 두 가지 주문 방식 지원: 일부 선택 주문 / 전체 주문

+++ src/main/java/com/hhplus/hhplus_ecommerce/cart/dto/response/CartDto.java
@@ -18,8 +18,8 @@ public record CartDto (
      Integer quantity,
 
     @Schema(description = "상품 단가", example = "2000000")
-     Integer price,
+     Long price,
 
     @Schema(description = "소계", example = "4000000")
-     Integer subtotal
+     Long subtotal
 ){}
