+++ src/main/java/com/hhplus/hhplus_ecommerce/cart/dto/response/CartListResponse.java
@@ -10,5 +10,5 @@ public record CartListResponse (
          List<CartDto> cartItems,
 
         @Schema(description = "총 금액", example = "5200000")
-         Integer totalAmount
+         Long totalAmount
 ){}
