+++ src/main/java/com/hhplus/hhplus_ecommerce/order/dto/response/OrderListDto.java
@@ -2,8 +2,6 @@ package com.hhplus.hhplus_ecommerce.order.dto.response;
 
 import com.hhplus.hhplus_ecommerce.order.OrderStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 import java.time.LocalDateTime;
 
@@ -14,7 +12,7 @@ public record OrderListDto (
      Long orderId,
 
     @Schema(description = "최종 금액", example = "50000")
-     Integer finalAmount,
+     Long finalAmount,
 
     @Schema(description = "주문 상태", example = "COMPLETED")
      OrderStatus status,
