+++ src/main/java/com/hhplus/hhplus_ecommerce/product/dto/response/PopularProductDto.java
@@ -1,8 +1,6 @@
 package com.hhplus.hhplus_ecommerce.product.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 @Schema(description = "인기 상품 정보")
 public record PopularProductDto (
@@ -13,7 +11,7 @@ public record PopularProductDto (
     String name,
 
     @Schema(description = "상품 가격", example = "2000000")
-    Integer price,
+    Long price,
 
     @Schema(description = "카테고리", example = "전자제품")
     String category,
